build shared library (#2525)
* build shared lib and enable lto

* reserved for layer and option

* allocator pimpl

* datareader pimpl

* paramdict pimpl, disable copy assign for allocator and datareader

* modelbin pimpl

* net extractor pimpl

* gpu pimple

* disable copy assign vulkandevice, code format

* command pimpl, dummy image readonly

* pipeline pipelinecache pimpl, export platform class

* code format, export simple family

* update ci

* disable lto on android armv7, merge webassembly ci

* link libgcc, fix macos dylib version

* pipeline pimpl, gpu info pimpl

* destroy gpu info after vulkan device

* ignore msvc stl class warning

* fix ncnn_paramdict_get_float return type

* fix vktransfer upload fp16 without flatten, add command test

+++ src/c_api.cpp
@@ -417,7 +417,7 @@ int ncnn_paramdict_get_int(const ncnn_paramdict_t pd, int id, int def)
     return ((const ParamDict*)pd)->get(id, def);
 }
 
-int ncnn_paramdict_get_float(const ncnn_paramdict_t pd, int id, float def)
+float ncnn_paramdict_get_float(const ncnn_paramdict_t pd, int id, float def)
 {
     return ((const ParamDict*)pd)->get(id, def);
 }
